5 service layer

+++ src/main/java/com/example/examensarbete/controller/UserController.java
@@ -1,42 +1,58 @@
 package com.example.examensarbete.controller;
 
-import com.example.examensarbete.dto.UserDto;
+import com.example.examensarbete.dto.GoogleUser;
+import com.example.examensarbete.service.AuthService;
 import com.example.examensarbete.service.UserService;
 import com.example.examensarbete.entities.User;
-import org.springframework.validation.annotation.Validated;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("api/v1/users")
 public class UserController {
     private final UserService userService;
+    private final AuthService authService;
 
-    public UserController(UserService userService){
+    public UserController(UserService userService,
+                          AuthService authService) {
         this.userService = userService;
+        this.authService = authService;
     }
 
-    /*
+    @GetMapping
+    public List<User> getAllUsers() {
+        return userService.getAllUsers();
+    }
+
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable int id){
+    public User getUserById(@PathVariable Long id) {
         return userService.getUserById(id);
     }
 
-    @PostMapping
-    public User addUser(@RequestBody @Validated UserDto user){
-        return userService.addUser(user);
-    }
-
-    @PatchMapping("/{id}")
-    public User editUser(@PathVariable int id, @RequestBody @Validated UserDto user){
-        return userService.editUser(id, user);
-    }
-
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable int id){
-        userService.deleteUser(id);
-        return "User with id: " + " was deleted.";
-    }
+    public ResponseEntity<?> deleteUser(@PathVariable Long id, Authentication authentication) {
+        if (authentication.getPrincipal() instanceof DefaultOAuth2User auth) {
+            GoogleUser googleUser = authService.getUserData(auth);
 
-     */
+            try {
+                if(userService.checkPermission(googleUser)){
+                    userService.deleteUser(id);
+                    return ResponseEntity.noContent().build();
+                }
+                return ResponseEntity.status(403).body("No permission");
+
+            } catch (RuntimeException e) {
+                return ResponseEntity.status(403).body(e.getMessage());
+            } catch (Exception e) {
+                return ResponseEntity.status(500).body("Internal Server Error");
+            }
+        }
+
+        return ResponseEntity.status(403).build();
+    }
 
 }
